Remove auto-createElement wrapper
Since events cannot be easily automated, remove the automatic wrapper creation via patching createElement.
