+++ packages/labs/rollup.config.js
@@ -15,10 +15,6 @@
 import {litProdConfig} from '../../rollup-common.js';
 
 export default litProdConfig({
-  entryPoints: [
-    'index',
-    'frameworks/react/create-component',
-    'frameworks/react/create-element',
-  ],
+  entryPoints: ['index', 'frameworks/react/create-component'],
   external: ['updating-element', 'lit-html', 'lit-element'],
 });
